Gradle build: add gradle build file.

+++ app/src/main/java/com/app/kimkulling/shoppinglist/SLDatabaseHelper.java
@@ -10,13 +10,19 @@ import android.util.Log;
  *  Helper class for SQLLite access.
  */
 public class SLDatabaseHelper extends SQLiteOpenHelper {
-    public static final String TAG       = "SLDatabaseHelper";
-    public static final String DB_NAME   = "shoppinglist.db";
-    public static final String SHL_TABLE = "shoppinglists";
-    public static final String C_ID      = BaseColumns._ID;
-    public static final String C_SHOP    = "shop";
-    public static final String C_LIST    = "item";
-    public static final int DB_VERSION   = 1;
+    public static final String TAG          = "SLDatabaseHelper";
+    public static final String DB_NAME      = "shoppinglist.db";
+    public static final String SHL_TABLE    = "shoppinglists";
+    public static final String C_SH_ID = BaseColumns._ID;
+    public static final String C_SH_SHOP = "shop";
+    public static final String C_SH_LIST = "item";
+    public static final String C_SH_CATID = "catid";
+    public static final String C_SH_ISIN = "isin";
+    public static final String C_SH_NUMITEMS = "number";
+    public static final String C_SHCAT_TABLE = "shopping_categories";
+    public static final String C_SHCAT_NAME = "category";
+
+    public static final int DB_VERSION   = 2;
 
     private final Context mCtx;
 
@@ -30,14 +36,19 @@ public class SLDatabaseHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * The onCreate-callback, will initially create the table
+     * The onCreate-callback, will initially create the tables.
      * @param db    The database instance
      */
     @Override
     public void onCreate( SQLiteDatabase db ) {
-        String sql = "create table " + SHL_TABLE + " (" + C_ID + " int primary key, "
-                + C_SHOP + " text, " + C_LIST + " text )";
-        executeSQL(db, sql);
+        String slSql = "create table " + SHL_TABLE + " (" + C_SH_ID + " int primary key, "
+                + C_SH_SHOP + " text, " + C_SH_LIST + " text, " + C_SH_CATID + " int )";
+        executeSQL( db, slSql );
+
+        String catSql = "create table " + C_SHCAT_TABLE + " (" + C_SH_ID + " int primary key, "
+                + C_SHCAT_NAME + " text )";
+        executeSQL( db, catSql );
+
         Log.d(TAG, " Create database " + DB_NAME);
     }
 
@@ -49,10 +60,12 @@ public class SLDatabaseHelper extends SQLiteOpenHelper {
      */
     @Override
     public void onUpgrade( SQLiteDatabase db, int oldVersion, int newVersion ) {
-        String sql = "drop table if exists " + SHL_TABLE;
-        executeSQL( db, sql );
+        if ( 1 == oldVersion ) {
+            String sql = "drop table if exists " + SHL_TABLE;
+            executeSQL(db, sql);
+            Log.d(TAG, " Update database " + DB_NAME + ", old database will be removed.");
+        }
 
-        Log.d(TAG, " Update database " + DB_NAME);
         onCreate( db );
     }
